back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.0",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "daphne==4.1.2",
    "django==5.1.3",
    "django-environ==0.11.2",
    "django-filter==24.3",
    "django-redis==5.4.0",
    "django-simpleui==2024.10.25",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.27.2",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "lxml>=6.0.2",
    "markdown==3.7",
    "mysqlclient>=2.2.7",
    "wechatpayv3==1.3.5",
    "wechatpy==1.8.18",
    "zhipuai>=2.1.5.20250825",
]
